examples/backend-python/requirements.txt
flask>=3.0
maarifx>=0.1.0
python-dotenv>=1.0
